Renaming output folders, removing sklearn logic to separate files
