+++ nodetypes/mltosca.nodes.Environment/Conda/NodeType.tosca
@@ -2,7 +2,7 @@ tosca_definitions_version: tosca_simple_yaml_1_3
 
 node_types:
   mltosca.nodes.Environment.Conda:
-    derived_from: tosca.nodes.SoftwareComponent
+    derived_from: mltosca.nodes.Environment
     metadata:
       targetNamespace: "mltosca.nodes.Environment"
       abstract: "false"
@@ -10,11 +10,6 @@ node_types:
     attributes:
       project_location:
         type: string
-    requirements:
-      - host:
-          capability: tosca.capabilities.Compute
-          relationship: tosca.relationships.HostedOn
-          occurrences: [ 1, 1 ]
     capabilities:
       host:
         occurrences: [ 1, UNBOUNDED ]
