+++ nodetypes/mltosca.nodes.Preprocessing/ReplaceNanValues/NodeType.tosca
@@ -11,6 +11,9 @@ node_types:
       function_id:
         type: string
         default: "replace_nan_values"
+      output_folder:
+        type: string
+        default: { get_attribute: [ SELF, host, output_folder ] }
       parameters:
         type: string
         default: { concat: [ { get_property: [ SELF, column ] }, "-|-", { get_property: [ SELF, method ] }, "-|-", { get_property: [ SELF, value ] } ] }
