+++ nodetypes/mltosca.nodes.Preprocessing/SortByColumn/NodeType.tosca
@@ -11,6 +11,9 @@ node_types:
       function_id:
         type: string
         default: "sort_by_column"
+      output_folder:
+        type: string
+        default: { get_attribute: [ SELF, host, output_folder ] }
       parameters:
         type: string
         default: { concat: [ { get_property: [ SELF, column ] }, "-|-", { get_property: [ SELF, order ] } ] }
