+++ nodetypes/mltosca.nodes.Sklearn/RandomForestClassifier/files/create/create.yml.mimetype
@@ -1,0 +1,1 @@
+text/plain
