+++ nodetypes/mltosca.nodes.Sklearn/RandomForestClassifier/files/script/script.py.mimetype
@@ -1,0 +1,1 @@
+text/x-python
