+++ nodetypes/mltosca.nodes.Splitting/SplitTrainTest/NodeType.tosca
@@ -10,7 +10,7 @@ node_types:
     attributes:
       function_id:
         type: string
-        default: "split_train_test"
+        default: { get_attribute: [ SELF, tosca_id ] }
     properties:
       test_size:
         type: float
