+++ nodetypes/mltosca.nodes.Splitting/SplitTrainTest/files/script/split.py
@@ -8,7 +8,7 @@ from pandas import DataFrame
 
 def setup_config_file(project_location, train_dataframe='train.pkl', test_dataframe='test.pkl'):
   with open(project_location + '/configuration.json', 'w') as file:
-    json.dump({'filenames': [train_dataframe, test_dataframe]}, file, indent=4)
+    json.dump({'filenames': [project_location + '/' + train_dataframe, project_location + '/' + test_dataframe]}, file, indent=4)
 
 
 def read_config_file(config_path) -> str:
@@ -37,9 +37,9 @@ if __name__ == "__main__":
     time.sleep(1)
   dataframe_file = read_config_file(previous_output_folder + '/configuration.json')
 
-  while not os.path.isfile(previous_output_folder + '/' +dataframe_file):
+  while not os.path.isfile(dataframe_file):
     time.sleep(1)
-  df = read_dataframe(previous_output_folder + '/' +dataframe_file)
+  df = read_dataframe(dataframe_file)
   train, test = train_test_split(df, test_size)
 
   train.to_pickle(output_folder + '/processing.pkl')
