+++ nodetypes/mltosca.nodes/Sklearn/NodeType.tosca
@@ -7,50 +7,6 @@ node_types:
       targetNamespace: "mltosca.nodes"
       abstract: "false"
       final: "false"
-    attributes:
-      output_folder:
-        type: string
-        default: { get_attribute: [ SELF, function_id ] }
-      parameters:
-        type: string
-        default: "none"
     properties:
       target:
         type: string
-    interfaces:
-      Standard:
-        type: tosca.interfaces.node.lifecycle.Standard
-        operations:
-          create:
-            description: The standard create operation
-            inputs:
-              data_folder:
-                type: string
-                required: true
-                default: { get_attribute: [ SELF, data_folder ] }
-              output_folder:
-                type: string
-                required: true
-                default: { get_attribute: [ SELF, function_id ] }
-              project_location:
-                type: string
-                required: false
-                default: { get_attribute: [ SELF, project_location ] }
-              parameters:
-                type: string
-                required: false
-                default: { get_attribute: [ SELF, parameters ] }
-              target:
-                type: string
-                required: true
-                default: { get_property: [ SELF, target ] }
-            implementation:
-              primary: create
-              timeout: 0
-    artifacts:
-      create:
-        type: radon.artifacts.Ansible
-        file: create.yml
-      script:
-        type: tosca.artifacts.File
-        file: script.py
