+++ relationshiptypes/mltosca.relationships/HostedOnPipeline/RelationshipType.tosca
@@ -23,6 +23,10 @@ relationship_types:
                 type: string
                 required: false
                 default: { get_attribute: [ SOURCE, project_location ] }
+              output_folder:
+                type: string
+                required: false
+                default: { get_attribute: [ TARGET, output_folder ] }
             implementation:
               primary: create.yml
               timeout: 0
