+++ servicetemplates/radon.blueprints.testing/test_simple/ServiceTemplate.tosca
@@ -25,16 +25,6 @@ topology_template:
             node: Conda_0
             relationship: con_HostedOn_3
             capability: host
-    RemoteWorkstation_0:
-      type: radon.nodes.VM.RemoteWorkstation
-      metadata:
-        x: "440"
-        y: "855"
-        displayName: "RemoteWorkstation"
-      properties:
-        ssh_username: "centos"
-        KeyFile: "/home/artjom/.ssh/id_rsa"
-        external_ip: "172.17.90.27"
     Precision_0:
       type: mltosca.nodes.Evaluation.Precision
       metadata:
@@ -204,11 +194,6 @@ topology_template:
         x: "424"
         y: "562"
         displayName: "Conda"
-      requirements:
-        - host:
-            node: RemoteWorkstation_0
-            relationship: con_HostedOn_6
-            capability: host
     ReadCSV_0:
       type: mltosca.nodes.Reading.ReadCSV
       metadata:
@@ -216,7 +201,7 @@ topology_template:
         y: "321"
         displayName: "ReadCSV"
       properties:
-        filepath: "/home/centos/data.csv"
+        filepath: "/home/artjom/thesis/TOSCA-ML/project1/data/data.csv"
         delimiter: ","
         encoding: "UTF-8"
       requirements:
@@ -255,26 +240,6 @@ topology_template:
             relationship: con_HostedOn_4
             capability: host
   relationship_templates:
-    con_DependsOn_0:
-      type: tosca.relationships.DependsOn
-    con_DependsOn_1:
-      type: tosca.relationships.DependsOn
-    con_DependsOn_4:
-      type: tosca.relationships.DependsOn
-    con_DependsOn_5:
-      type: tosca.relationships.DependsOn
-    con_DependsOn_2:
-      type: tosca.relationships.DependsOn
-    con_DependsOn_3:
-      type: tosca.relationships.DependsOn
-    con_HostedOnEvaluation_1:
-      type: mltosca.relationships.HostedOnEvaluation
-    con_HostedOnEvaluation_0:
-      type: mltosca.relationships.HostedOnEvaluation
-    con_HostedOnEvaluation_3:
-      type: mltosca.relationships.HostedOnEvaluation
-    con_HostedOnEvaluation_2:
-      type: mltosca.relationships.HostedOnEvaluation
     con_HostedOnPipeline_4:
       type: mltosca.relationships.HostedOnPipeline
     con_HostedOn_2:
@@ -297,8 +262,6 @@ topology_template:
       type: tosca.relationships.DependsOn
     con_DependsOn_7:
       type: tosca.relationships.DependsOn
-    con_HostedOn_6:
-      type: tosca.relationships.HostedOn
     con_HostedOnPipeline_0:
       type: mltosca.relationships.HostedOnPipeline
     con_HostedOnPipeline_1:
@@ -307,3 +270,23 @@ topology_template:
       type: tosca.relationships.HostedOn
     con_HostedOn_5:
       type: tosca.relationships.HostedOn
+    con_DependsOn_0:
+      type: tosca.relationships.DependsOn
+    con_DependsOn_1:
+      type: tosca.relationships.DependsOn
+    con_DependsOn_4:
+      type: tosca.relationships.DependsOn
+    con_DependsOn_5:
+      type: tosca.relationships.DependsOn
+    con_DependsOn_2:
+      type: tosca.relationships.DependsOn
+    con_DependsOn_3:
+      type: tosca.relationships.DependsOn
+    con_HostedOnEvaluation_1:
+      type: mltosca.relationships.HostedOnEvaluation
+    con_HostedOnEvaluation_0:
+      type: mltosca.relationships.HostedOnEvaluation
+    con_HostedOnEvaluation_3:
+      type: mltosca.relationships.HostedOnEvaluation
+    con_HostedOnEvaluation_2:
+      type: mltosca.relationships.HostedOnEvaluation
